feat: [Article] contentView를 분리해요.
